chore(refactor): return typed args from cli parser
The cli parsed used to return an 'argparse.Namespace' which is basically just
a dictionary. Hence, there was no real possibility to statically assert typing.

By introducing NamedTuple args for every command we have a fully typed interface
for the cli parser. The type checks between the argparse.Namespace and the new
NamedTuples are done during unit testing by either the NamedTuple constructor or
typeguard.