feat: Pull and rebase before push
Resolves #229

If another user pushes changes to the remote repo after gitopscli has cloned it,
the subsequent push will fail. Alliviate this problem by pulling and rebasing
local changes on top of any remote changes just before pushing.:

For the case of creating a new branch (for a PR), the pull is skipped because
it would fail due to the missing remote branch.